YAY, now floating point as the internal audio format is supported.  It can be enabled with a configure flag Usually samples are converted with a template function, convert_sample<from_type,to_type>() that is specialized for all supported conversions, more can easily be added More work still needs to be done--namely, making .rez support more than just 16bit PCM format and making COSSSoundPlayer initialize for something better than 16bit trying several formats

+++ rezound/src/backend/Effects/CDistortionEffect.cpp
@@ -39,7 +39,8 @@ bool CDistortionEffect::doActionSizeSafe(CActionSound &actionSound,bool prepareF
 			sample_pos_t destPos=start;
 			CRezPoolAccesser dest=actionSound.sound->getAudio(i);
 			
-			TDSPDistorter<sample_t,MAX_SAMPLE> d(curve,curve);
+						//??? wouldn't work well if it wasn't an integral value, perhaps I should make it a parameter to the constructor
+			TDSPDistorter<sample_t,(int)MAX_SAMPLE> d(curve,curve);
 
 			while(destPos<=stop)
 			{
